requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
marshmallow==3.20.1
pytest==7.3.2